Add ffmpeg support for Ubuntu 20.04 and switch to python 3.8 for tests (#1195)
* Add ffmpeg support for Ubuntu 20.04

In the past only Ubuntu 16.04 and 18.04 are supported for ffmpeg audio/video
Dataset.

In this PR ffmpeg support for Ubuntu 20.04 has been added as Ubuntu 20.04 has
been released this year.

Also fix `Invalid return value 0 for stream protocol` issue when LIBAVFORMAT_VERSION_MAJOR > 57

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Switch to python 3.8 for tests

In the past we use python 3.6/3.7 as the default version for testing
as some packages for testing are not available for 3.8 yet,
even though many python systems have been upgraded to 3.8. For example,
Ubuntu 20.04 and macOS's default system pythons are already 3.8.

This PR switches to use python 3.8 for testings. This should also reduce
GitHub Actions jobs a little bit.

This PR also updates ffmpeg video samples (as there is a small different between
Ubuntu 18.04 and Ubuntu 20.04 ffmpeg)

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add python3.7 test back, based on review comment

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>