Convert to use tensorflow's third party BUILD files whenever possible (#1395)
* Remove unneeded libraries

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Convert to use tensorflow's third party BUILD files whenever possible

In tensorflow-io, while there are many third-party libraries such as
arrow/avro that was not used by tensorflow, quite a few libraries such
as boringssl/curl/nasm are used by tensorflow.

In the past we used to maintain our own BUILD files for third party libraries
such as curl. Which potentially causes duplication (with tensorflow), and may
cause version incompatibility if we are out of sync with tensorflow's third party
libraries.

This PR uses tensorflow's BUILD files whenever possible so that we can reduce
overlap.

Note several libraries needs special handling in tensorflow-io, e.g., snappy
in tensorflow-io needs inclusion of snappy-c. We can see if it is possible to
update upstream tensorflow instead.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Windows fix

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Windows fix

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Try fix

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Fix

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Emtpy commit to trigger CI

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Bump to tensorflow 2.5.0

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add bcrypt as dependency in Windows

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Remove unneeded python setup

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>